+++ app.py
@@ -12,5 +12,5 @@ print("SystemInfo -> ", systemInfo)
 # print(pytz.all_timezones) # List out all the timezone available
 if __name__ == '__main__':
     app = create_app()
-    app.run(port=int(en_var("PORT", 8080)),
-            debug=en_var("DEBUG"))  # , host='0.0.0.0',
+    # app.run(port=int(en_var("PORT", 8080)),
+    #         debug=en_var("DEBUG", False))  # , host='0.0.0.0',